==> analysis/05-gene-ontology.Rmd <==
---
title: "Gene ontology analysis"
author: "James Ashmore"
date: "2020-04-22"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Setup

Set chunk options:

```{r}
knitr::opts_chunk$set(
  autodep    = TRUE,
  cache      = TRUE,
  cache.path = "cache/05-gene-ontology.Rmd/",
  dev        = "png",
  error      = FALSE,
  message    = FALSE,
  warning    = FALSE
)
```

Load required packages:

```{r}
pkg <- c("limma", "org.Mm.eg.db", "scater", "scran")

lib <- lapply(pkg, library, character.only = TRUE)
```

Import experiment data:

```{r}
sce <- readRDS("data/04-differential-analysis.rds")
```

Define a set of background universe genes:

```{r}
dec <- modelGeneVar(sce, block = sce$batch)

hvg <- getTopHVGs(dec, var.threshold = 0) # take all genes with biological variation

ens <- rowData(sce)$gene_id[match(hvg, rownames(sce))]

uni <- mapIds(org.Mm.eg.db, keys = ens, column = "ENTREZID", keytype = "ENSEMBL")
```

Extract gene identifiers with the cell surface gene ontology term:

```{r}
ann <- select(org.Mm.eg.db, keys = "GO:0009986", columns = "ENSEMBL", keytype = "GO")
```

## Helper functions

Define wrapper for the goana function:

```{r}
calculateGO <- function(query, subject) {
  
  res <- goana(de = query, universe = subject, species = "Mm")
  
  res <- split(res, res$Ont)
  
  res <- lapply(res, function(x) x[order(x$P.DE, decreasing = FALSE), ])
  
  res <- lapply(res, function(x) { x$FDR <- p.adjust(x$P.DE, method = "fdr"); x} )
  
  res <- lapply(res, subset, FDR < 0.05)
  
  return(res)

}
```

Define function to visualise GO results:

```{r}
visualiseGO <- function(results, n = 10) {
  
  results <- head(results, n = n)
  
  if (nrow(results) == 0) { return(ggplot() + theme_void()) }
  
  results$Term <- tools::toTitleCase(results$Term)
  
  results$Ratio <- results$DE / results$N

  plt <- ggplot(results, aes(x = Ratio, y = reorder(Term, Ratio), colour = -log10(FDR), size = DE)) + 
    geom_point() + 
    scale_colour_viridis_c() + 
    labs(x = "Ratio", size = "Count", colour = "-log10(FDR)") + 
    guides(colour = guide_colorbar(order = 1), size = guide_legend(order = 2)) +
    theme_bw() + 
    theme(axis.title.y = element_blank())
  
  return(plt)

}
```

Define wrapper for the kegga function:

```{r}
calculateKEGG <- function(query, subject) {
  
  res <- kegga(de = query, universe = subject, species = "Mm")

  res <- res[order(res$P.DE, decreasing = FALSE), ]

  res$FDR <- p.adjust(res$P.DE, method = "fdr")
  
  res <- sig <- subset(res, FDR < 0.05)
  
  return(res)

}
```

Define function to visualise KEGG results:

```{r}
visualiseKEGG <- function(results, n = 10) {
  
  results <- head(results, n = n)
  
  if (nrow(results) == 0) { return(ggplot() + theme_void()) }
  
  results$Pathway <- tools::toTitleCase(results$Pathway)
  
  results$Ratio <- results$DE / results$N

  plt <- ggplot(results, aes(x = Ratio, y = reorder(Pathway, Ratio), colour = -log10(FDR), size = DE)) + 
    geom_point() + 
    scale_colour_viridis_c() + 
    labs(x = "Ratio", size = "Count", colour = "-log10(FDR)") + 
    guides(colour = guide_colorbar(order = 1), size = guide_legend(order = 2)) +
    theme_bw() + 
    theme(axis.title.y = element_blank())
  
  return(plt)

}
```

## Correlation Test

### Lineage 1

Extract and convert gene identifiers from results table:

```{r}
dat <- metadata(sce)$correlateTime$slingPseudotime_1

ens <- rowData(sce)$gene_id[match(dat$gene2, rownames(sce))]

ids <- mapIds(org.Mm.eg.db, keys = ens, column = "ENTREZID", keytype = "ENSEMBL")
```

Print results table with cell surface annotation:

```{r}
dat$ontology <- ifelse(ens %in% ann$ENSEMBL, "cell surface", NA)

DT::datatable(as.data.frame(dat))
```

Test for over-representation of gene ontology terms and KEGG pathways:

```{r}
res <- calculateGO(query = ids, subject = uni)

res$KEGG <- calculateKEGG(query = ids, subject = uni)
```

Plot the 10 most significant gene ontology terms and KEGG pathways:

```{r fig.height = 14, fig.width = 9}
plt <- list(
  visualiseGO(res$BP, n = 10) + ggtitle("Biological Process"),
  visualiseGO(res$CC, n = 10) + ggtitle("Cellular Component"),
  visualiseGO(res$MF, n = 10) + ggtitle("Molecular Function"),
  visualiseKEGG(res$KEGG, n = 10) + ggtitle("KEGG Pathway")
)

patchwork::wrap_plots(plt, ncol = 1)
```

### Lineage 2

Extract and convert gene identifiers from results table:

```{r}
dat <- metadata(sce)$correlateTime$slingPseudotime_2

ens <- rowData(sce)$gene_id[match(dat$gene2, rownames(sce))]

ids <- mapIds(org.Mm.eg.db, keys = ens, column = "ENTREZID", keytype = "ENSEMBL")
```

Print results table with cell surface annotation:

```{r}
dat$ontology <- ifelse(ens %in% ann$ENSEMBL, "cell surface", NA)

DT::datatable(as.data.frame(dat))
```

Test for over-representation of gene ontology terms and KEGG pathways:

```{r}
res <- calculateGO(query = ids, subject = uni)

res$KEGG <- calculateKEGG(query = ids, subject = uni)
```

Plot the 10 most significant gene ontology terms and KEGG pathways:

```{r fig.height = 14, fig.width = 9}
plt <- list(
  visualiseGO(res$BP, n = 10) + ggtitle("Biological Process"),
  visualiseGO(res$CC, n = 10) + ggtitle("Cellular Component"),
  visualiseGO(res$MF, n = 10) + ggtitle("Molecular Function"),
  visualiseKEGG(res$KEGG, n = 10) + ggtitle("KEGG Pathway")
)

patchwork::wrap_plots(plt, ncol = 1)
```

## Association Test

### Lineage 1

Extract and convert gene identifiers from results table:

```{r}
dat <- metadata(sce)$associationTest$slingPseudotime_1

ens <- rowData(sce)$gene_id[match(rownames(dat), rownames(sce))]

ids <- mapIds(org.Mm.eg.db, keys = ens, column = "ENTREZID", keytype = "ENSEMBL")
```

Print results table with cell surface annotation:

```{r}
dat$ontology <- ifelse(ens %in% ann$ENSEMBL, "cell surface", NA)

DT::datatable(as.data.frame(dat))
```

Test for over-representation of gene ontology terms and KEGG pathways:

```{r}
res <- calculateGO(query = ids, subject = uni)

res$KEGG <- calculateKEGG(query = ids, subject = uni)
```

Plot the 10 most significant gene ontology terms and KEGG pathways:

```{r fig.height = 14, fig.width = 9}
plt <- list(
  visualiseGO(res$BP, n = 10) + ggtitle("Biological Process"),
  visualiseGO(res$CC, n = 10) + ggtitle("Cellular Component"),
  visualiseGO(res$MF, n = 10) + ggtitle("Molecular Function"),
  visualiseKEGG(res$KEGG, n = 10) + ggtitle("KEGG Pathway")
)

patchwork::wrap_plots(plt, ncol = 1)
```

### Lineage 2

Extract and convert gene identifiers from results table:

```{r}
dat <- metadata(sce)$associationTest$slingPseudotime_2

ens <- rowData(sce)$gene_id[match(rownames(dat), rownames(sce))]

ids <- mapIds(org.Mm.eg.db, keys = ens, column = "ENTREZID", keytype = "ENSEMBL")
```

Print results table with cell surface annotation:

```{r}
dat$ontology <- ifelse(ens %in% ann$ENSEMBL, "cell surface", NA)

DT::datatable(as.data.frame(dat))
```

Test for over-representation of gene ontology terms and KEGG pathways:

```{r}
res <- calculateGO(query = ids, subject = uni)

res$KEGG <- calculateKEGG(query = ids, subject = uni)
```

Plot the 10 most significant gene ontology terms and KEGG pathways:

```{r fig.height = 14, fig.width = 9}
plt <- list(
  visualiseGO(res$BP, n = 10) + ggtitle("Biological Process"),
  visualiseGO(res$CC, n = 10) + ggtitle("Cellular Component"),
  visualiseGO(res$MF, n = 10) + ggtitle("Molecular Function"),
  visualiseKEGG(res$KEGG, n = 10) + ggtitle("KEGG Pathway")
)

patchwork::wrap_plots(plt, ncol = 1)
```

## Start vs End Test

### Lineage 1

Extract and convert gene identifiers from results table:

```{r}
dat <- metadata(sce)$startVsEndTest$slingPseudotime_1

ens <- rowData(sce)$gene_id[match(rownames(dat), rownames(sce))]

ids <- mapIds(org.Mm.eg.db, keys = ens, column = "ENTREZID", keytype = "ENSEMBL")
```

Print results table with cell surface annotation:

```{r}
dat$ontology <- ifelse(ens %in% ann$ENSEMBL, "cell surface", NA)

DT::datatable(as.data.frame(dat))
```

Test for over-representation of gene ontology terms and KEGG pathways:

```{r}
res <- calculateGO(query = ids, subject = uni)

res$KEGG <- calculateKEGG(query = ids, subject = uni)
```

Plot the 10 most significant gene ontology terms and KEGG pathways:

```{r fig.height = 14, fig.width = 9}
plt <- list(
  visualiseGO(res$BP, n = 10) + ggtitle("Biological Process"),
  visualiseGO(res$CC, n = 10) + ggtitle("Cellular Component"),
  visualiseGO(res$MF, n = 10) + ggtitle("Molecular Function"),
  visualiseKEGG(res$KEGG, n = 10) + ggtitle("KEGG Pathway")
)

patchwork::wrap_plots(plt, ncol = 1)
```

### Lineage 2

Extract and convert gene identifiers from results table:

```{r}
dat <- metadata(sce)$startVsEndTest$slingPseudotime_2

ens <- rowData(sce)$gene_id[match(rownames(dat), rownames(sce))]

ids <- mapIds(org.Mm.eg.db, keys = ens, column = "ENTREZID", keytype = "ENSEMBL")
```

Print results table with cell surface annotation:

```{r}
dat$ontology <- ifelse(ens %in% ann$ENSEMBL, "cell surface", NA)

DT::datatable(as.data.frame(dat))
```

Test for over-representation of gene ontology terms and KEGG pathways:

```{r}
res <- calculateGO(query = ids, subject = uni)

res$KEGG <- calculateKEGG(query = ids, subject = uni)
```

Plot the 10 most significant gene ontology terms and KEGG pathways:

```{r fig.height = 14, fig.width = 9}
plt <- list(
  visualiseGO(res$BP, n = 10) + ggtitle("Biological Process"),
  visualiseGO(res$CC, n = 10) + ggtitle("Cellular Component"),
  visualiseGO(res$MF, n = 10) + ggtitle("Molecular Function"),
  visualiseKEGG(res$KEGG, n = 10) + ggtitle("KEGG Pathway")
)

patchwork::wrap_plots(plt, ncol = 1)
```

## diffEndTest

Extract and convert gene identifiers from results table:

```{r}
dat <- metadata(sce)$diffEndTest

ens <- rowData(sce)$gene_id[match(rownames(dat), rownames(sce))]

ids <- mapIds(org.Mm.eg.db, keys = ens, column = "ENTREZID", keytype = "ENSEMBL")
```

Print results table with cell surface annotation:

```{r}
dat$ontology <- ifelse(ens %in% ann$ENSEMBL, "cell surface", NA)

DT::datatable(as.data.frame(dat))
```

Test for over-representation of gene ontology terms and KEGG pathways:

```{r}
res <- calculateGO(query = ids, subject = uni)

res$KEGG <- calculateKEGG(query = ids, subject = uni)
```

Plot the 10 most significant gene ontology terms and KEGG pathways:

```{r fig.height = 14, fig.width = 9}
plt <- list(
  visualiseGO(res$BP, n = 10) + ggtitle("Biological Process"),
  visualiseGO(res$CC, n = 10) + ggtitle("Cellular Component"),
  visualiseGO(res$MF, n = 10) + ggtitle("Molecular Function"),
  visualiseKEGG(res$KEGG, n = 10) + ggtitle("KEGG Pathway")
)

patchwork::wrap_plots(plt, ncol = 1)
```

## Pattern Test

Extract and convert gene identifiers from results table:

```{r}
dat <- metadata(sce)$patternTest

ens <- rowData(sce)$gene_id[match(rownames(dat), rownames(sce))]

ids <- mapIds(org.Mm.eg.db, keys = ens, column = "ENTREZID", keytype = "ENSEMBL")
```

Print results table with cell surface annotation:

```{r}
dat$ontology <- ifelse(ens %in% ann$ENSEMBL, "cell surface", NA)

DT::datatable(as.data.frame(dat))
```

Test for over-representation of gene ontology terms and KEGG pathways:

```{r}
res <- calculateGO(query = ids, subject = uni)

res$KEGG <- calculateKEGG(query = ids, subject = uni)
```

Plot the 10 most significant gene ontology terms and KEGG pathways:

```{r fig.height = 14, fig.width = 9}
plt <- list(
  visualiseGO(res$BP, n = 10) + ggtitle("Biological Process"),
  visualiseGO(res$CC, n = 10) + ggtitle("Cellular Component"),
  visualiseGO(res$MF, n = 10) + ggtitle("Molecular Function"),
  visualiseKEGG(res$KEGG, n = 10) + ggtitle("KEGG Pathway")
)

patchwork::wrap_plots(plt, ncol = 1)
```

## Early DE Test

Extract and convert gene identifiers from results table:

```{r}
dat <- metadata(sce)$earlyDETest

ens <- rowData(sce)$gene_id[match(rownames(dat), rownames(sce))]

ids <- mapIds(org.Mm.eg.db, keys = ens, column = "ENTREZID", keytype = "ENSEMBL")
```

Print results table with cell surface annotation:

```{r}
dat$ontology <- ifelse(ens %in% ann$ENSEMBL, "cell surface", NA)

DT::datatable(as.data.frame(dat))
```

Test for over-representation of gene ontology terms and KEGG pathways:

```{r}
res <- calculateGO(query = ids, subject = uni)

res$KEGG <- calculateKEGG(query = ids, subject = uni)
```

Plot the 10 most significant gene ontology terms and KEGG pathways:

```{r fig.height = 14, fig.width = 9}
plt <- list(
  visualiseGO(res$BP, n = 10) + ggtitle("Biological Process"),
  visualiseGO(res$CC, n = 10) + ggtitle("Cellular Component"),
  visualiseGO(res$MF, n = 10) + ggtitle("Molecular Function"),
  visualiseKEGG(res$KEGG, n = 10) + ggtitle("KEGG Pathway")
)

patchwork::wrap_plots(plt, ncol = 1)
```
